Inline trait CanBeInNavigation into AbstractPage
